Feat (BE-048): refactor IT and extract to unit test if possible
